# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "A lightweight, containerized REST API service that provides video and audio download and streaming functionalities using yt-dlp and ffmpeg.",
        "title": "GoStreamPuller API",
        "contact": {
            "name": "API Support",
            "url": "http://www.example.com/support",
            "email": "support@example.com"
        },
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/stream/audio": {
            "post": {
                "description": "Streams an audio file directly from the source URL.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "audio/mpeg"
                ],
                "tags": [
                    "stream"
                ],
                "summary": "Stream an audio file",
                "parameters": [
                    {
                        "description": "Audio stream request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.StreamAudioRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully streamed audio",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "400": {
                        "description": "Invalid request payload or missing URL",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error during audio streaming",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/stream/video": {
            "post": {
                "description": "Streams a video directly from the source URL.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "video/mp4"
                ],
                "tags": [
                    "stream"
                ],
                "summary": "Stream a video",
                "parameters": [
                    {
                        "description": "Video stream request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.StreamVideoRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully streamed video",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "400": {
                        "description": "Invalid request payload or missing URL",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error during video streaming",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "handler.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "handler.StreamAudioRequest": {
            "type": "object",
            "properties": {
                "bitrate": {
                    "type": "string"
                },
                "codec": {
                    "type": "string"
                },
                "outputFormat": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "handler.StreamVideoRequest": {
            "type": "object",
            "properties": {
                "codec": {
                    "type": "string"
                },
                "format": {
                    "type": "string"
                },
                "resolution": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
...
    }
}